refactoring & cleanup in ts tooling
