+++ ethereum/package.json
@@ -21,8 +21,6 @@
     "integration-test": "bash shell-scripts/run_integration_tests.sh",
     "typechain": "bash ../sdk/scripts/make_ethers_types.sh",
     "flatten": "mkdir -p node_modules/@poanet/solidity-flattener/contracts && cp -r contracts/* node_modules/@poanet/solidity-flattener/contracts/ && poa-solidity-flattener",
-    "registerChains": "ts-node ./ts-scripts/registerChains.ts",
-    "messageTest": "ts-node ./ts-scripts/messageTest.ts",
     "deployAndConfigureTilt": "ENV=tilt bash ./ts-scripts/shell/deployConfigureTest.sh"
   },
   "author": "",
