+++ ethereum/ts-scripts/config/tilt/contracts.json
@@ -1,5 +1,5 @@
 {
-  "description": "This file contains the addresses for the relayProviderProxy and CoreRelayerProxy on each chain. If useLastRun is true, this file will be ignored, and the addresses will be taken from the lastrun.json of the deployment scripts.",
+  "description": "This file contains the addresses for the contracts on each chain. If useLastRun is true, this file will be ignored, and the addresses will be taken from the lastrun.json of the deployment scripts.",
   "useLastRun": true,
   "relayProviders": [
     {
