--- ethereum/ts-scripts/config/tilt/scriptConfigs/configureCoreRelayer.json
@@ -1,1 +1,0 @@
-{}
