--- ethereum/ts-scripts/config/tilt/scriptConfigs/deployCoreRelayer.json
@@ -1,1 +1,0 @@
-{}
