--- ethereum/ts-scripts/config/tilt/scriptConfigs/deployRelayProvider.json
@@ -1,1 +1,0 @@
-{}
