--- ethereum/ts-scripts/config/tilt/scriptConfigs/registerChainsCoreRelayer.json
@@ -1,1 +1,0 @@
-{}
