--- ethereum/ts-scripts/config/tilt/scriptConfigs/registerChainsRelayProvider.json
@@ -1,1 +1,0 @@
-{}
